+++ packages/babel-plugin/src/index.ts
@@ -1,3 +1,5 @@
+// @ts-nocheck
+
 import fs from 'fs'
 import path from 'path'
 import crypto from 'crypto'
@@ -9,6 +11,8 @@ import fg from 'fast-glob'
 import * as classNameStrategies from '@/classNameStrategies'
 import { isCCSSTag, covertToStringLiteralTag, getIdentifierByValueType, getAttrDetails } from '@/helpers'
 import { hybrid, onlyFullyStatic } from '@/handlers'
+
+export { hybrid, onlyFullyStatic } from '@/handlers'
 
 const defaultOpts = {
     identifiers: {
@@ -80,6 +84,9 @@ export default (api, opts) => {
             camelLong,
             processor: ccssOptions.props[camelShort]
         }
+        // Attach default handler
+        ccssPropMap[camelShort].processor.babelPluginHandler =
+            ccssPropMap[camelShort].processor.babelPluginHandler || ((...args) => onlyFullyStatic(...args, true))
     }
     // Add it for 3rd party props also (like mq, child, etc)
     for (const k of Object.keys(ccssOptions.props)) {
@@ -97,7 +104,7 @@ export default (api, opts) => {
         processor.babelPluginHandler = processor.babelPluginHandler || onlyFullyStatic
     }
     // Child needs static handler
-    ccssPropMap.child.processor.babelPluginHandler = hybrid
+    ccssPropMap.child.processor.babelPluginHandler = (attr, state, t, api) => hybrid(attr, state, t, api, true)
 
     opts.ccssPropMap = ccssPropMap