+++ packages/babel-plugin/src/test.tsx
@@ -2,13 +2,23 @@
  * @jest-environment node
  */
 
+// @ts-nocheck
+
 import pluginTester from 'babel-plugin-tester/pure'
-import plugin from '../'
+import plugin from '.'
+
+const options = require('ccss').defaultOptions
+
+// Custom prop
+options.props.button = () => ''
 
 pluginTester({
     pluginName: 'babel-plugin-ccss',
     plugin,
     pluginOptions: {
+        expressions: {
+            options
+        },
         identifiers: {
             Ui: true,
             DP: {
@@ -23,6 +33,21 @@ pluginTester({
                     Small: 10,
                     Extras: {
                         Foo: 'Bar'
+                    }
+                },
+                Button: {
+                    Size: {
+                        XSmall: 'x-small'
+                    }
+                },
+                Theme: {
+                    Foo: {
+                        Bar: {
+                            Baz: {
+                                Array: [0, 0, 0, '#000'],
+                                Object: { foo: 1 }
+                            }
+                        }
                     }
                 }
             },
@@ -268,6 +293,26 @@ pluginTester({
             title: 'handles defaultProps',
             code: `<DP />;`,
             output: `<div data-ui />;`
+        },
+        {
+            title: 'handles MemberExpressions',
+            code: `<Ui button={{ size: Ui.Button.Size.XSmall }} />;`,
+            output: `<div className="button___size___x_small__" />;`
+        },
+        {
+            title: 'handles non-ccss MemberExpressions',
+            code: `<Ui button={{ color: Ui.Button.Size.XSmall }} />;`,
+            output: `<div className="button___color___x_small__" />;`
+        },
+        {
+            title: 'handles prepared values => array',
+            code: `<Ui boxShadow={Ui.Theme.Foo.Bar.Baz.Array}>{variable}</Ui>;`,
+            output: `<div className="box_shadow__0_0_0___000__">{variable}</div>;`
+        },
+        {
+            title: 'handles prepared values => object',
+            code: `<Ui boxShadow={Ui.Theme.Foo.Bar.Baz.Object}>{variable}</Ui>;`,
+            output: `<div className="box_shadow___foo__1_">{variable}</div>;`
         }
     ]
 })