+++ packages/core/src/sandbox/jsx.tsx
@@ -4,9 +4,9 @@ import { ccss } from '../ccss'
 
 // Uncomment for testing only
 
-/*import ICCSSProps from '../../dist/types'
+/*import CCSSProps from '../../dist/types'
 
-const Cmp = styled.div<ICCSSProps>(ccss)
+const Cmp = styled.div<CCSSProps>(ccss)
 
 const App = () => {
     return <Cmp bgbm={} />