+++ packages/plugin-gutter/src/index.ts
@@ -1,4 +1,4 @@
-import { pipe } from '@cryptic-css/core'
+import { pipe, CCSSOptions, CCSSParser, CCSSProps } from '@cryptic-css/core'
 
 const DEFAULT = 5
 
@@ -17,20 +17,21 @@ const gutter = (input, prop, options) => {
     }
 }
 
-export default ({ props }) => {
-    Object.assign(props, {
-        p: pipe(gutter, props.p),
-        pT: pipe(gutter, props.pT),
-        pR: pipe(gutter, props.pR),
-        pB: pipe(gutter, props.pB),
-        pL: pipe(gutter, props.pL),
-        m: pipe(gutter, props.m),
-        mT: pipe(gutter, props.mT),
-        mR: pipe(gutter, props.mR),
-        mB: pipe(gutter, props.mB),
-        mL: pipe(gutter, props.mL),
-        gg: pipe(gutter, props.gg),
-        grg: pipe(gutter, props.grg),
-        gcg: pipe(gutter, props.gcg)
+export default (options: Partial<CCSSOptions>) => {
+    const props = options.props as CCSSProps
+    Object.assign(props as CCSSProps, {
+        p: pipe(gutter, props.p as CCSSParser),
+        pT: pipe(gutter, props.pT as CCSSParser),
+        pR: pipe(gutter, props.pR as CCSSParser),
+        pB: pipe(gutter, props.pB as CCSSParser),
+        pL: pipe(gutter, props.pL as CCSSParser),
+        m: pipe(gutter, props.m as CCSSParser),
+        mT: pipe(gutter, props.mT as CCSSParser),
+        mR: pipe(gutter, props.mR as CCSSParser),
+        mB: pipe(gutter, props.mB as CCSSParser),
+        mL: pipe(gutter, props.mL as CCSSParser),
+        gg: pipe(gutter, props.gg as CCSSParser),
+        grg: pipe(gutter, props.grg as CCSSParser),
+        gcg: pipe(gutter, props.gcg as CCSSParser)
     })
 }